release v0.10.1: slight tweak to markdown images
I've slightly tweaked how markdown images are exported, in that, I've
added a set height in pixels so that images have a consistent size.
I've chosen height as the oats database, is by design, oriented top
to bottom, so it just felt right to control that variable.
And, after some trial and error, I've found that 720 is the best height
for images as it allows for the image to both be small enough to fit
within half my screen
(as usually I have a split master layout on hyprland)
and also large enough to appreciate and view the details.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>