videos: added proper support for video formats
I've implemented simple support for the various video formats that html
supports (mp4, ogg & webm).
I've done this through a simple metadata enum for the 'is_vid'
(now renamed 'video_kind' for the feature) flag.

The video kind is also printed when exporting to normal format.

The video format is detected by simply reading the magic number of the
provided file (or stdin), returning an error if the video format is not
recognised (which is good validation).

Everything else should be pretty self-explanitory.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>