release v0.11.1: extended the count command
Frontend changes made:
  * added the ability to include certain attributes/features to count
  * added the ability to 'not' include certain attributes/features to
    count

In this release, I've extended the oats count command to also be able
to count how many of a feature is featured in the stack database.
This alone was pretty good, but I've also added an optional '--not' flag
that does the opposite of including an attribute, in that it won't count
attributes with it.

Also, this only counts as a patch change, as the 'major' & 'minor'
semantic version changes shall only apply to the oats database file
format.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>