v0.11.2: import binary search & markdown bugfixes
Changes made to the frontend:
  * binary search search no longer crashes when inserting to the start
    of a stack
  * added newlines to 'Included things' so they display correctly in
    markdown (when exported)

Binary search, I've noticed to crash due to there being either a (0 - 1)
error (when the searched stack is empty) or, if the item should be
inserted to the start of the stack, the 'high' must be lower than 'low'
and 'low' must be zero. So I've changed it so it checks for empty stacks
and also allows for high to be negative.

Markdown needs extra newlines when it's not a list item
(otherwise it gets included in the previous line)>

Signed-off-by: Ethan Ma <kalscium@protonmail.com>