v0.11.3: image, file & video pushing bugfix
I've recently come across a weird bug when I pushed images where some
images would be pushed with the same id to the database- leading to
dataloss when importing / exporting.
The reason, turned out to be that the code was time-dependant, as the
id it used was the current time in milliseconds, so, due to the
efficientcy of the program, if the file was sufficiently small, it would
be pushed before the millisecond was over, and the next would be pushed
with the same id.

I've solved this issue by simply introducing a time delay of 1ms so that
it takes at *least* that long and has an id change of at least 1.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>