release v0.12.0: added debug cmds & setid, settime
Frontend changes made:
  * added secret debug commands for recovering data
  * added the setid debug command to change the ids of an item
  * added the settime debug command to change the timestamp of an item

I've added support for debug commands, as, during the testing of oats, I
sometimes come across bugs and issues that might cause data-loss.

I've implemented the setid debug command, so I can change the ids of
items with duplicate ids, this command only affects the last item found
with this id (the latest) as the older stack items should take
precedence, and be harder to change.

I've also implemented the settime debug command, under the same idea
as the setid debug command, to be only used for data RECOVERY, and when
it is necessary to handle an incoherent database.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>