v0.12.1: void item markdown export bugfix
There was this issue, where items immedietely after a trimmed item, or
a collection of trimmed items would be skipped and not exported to
markdown.
This was due to the exporting of trimmed items advancing `i` to the next
item, then hitting continue which would skip it.
I've solved this simply by using the same solution I'd had for img & vid
exporting.

I've also added an extra newline at the start of trimmed item markdown
exports so that it displays properly in HTML
(and doesn't get wrapped to the last line).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>