v0.12.5: use-after-free bugfix for markdown export
Changes made:
  * (~) fixed features collection use after free error
  * (~) added features freeing to collection deinit
  * (~) added freeing for pushing stubbed items

I've encountered yet another memory management error in my code.
So, previously, I just did a search and replace to implement the freeing
of features with a defer.
However, the plan with the features collection, is that they would all
be loaded into memory into an array, and then freed at once at the end,
but because of the search and replace solution, I'd freed the feature
before inserting it into the collection, making it invalid from the
start.

Also, the collection freeing, didn't have support for freeing features.

Also, I forgot a `defer allocator.free` for pushing stubbed items after
packing them.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>