v0.12.6: another memory leak bugfix
Changes made to the frontend:
  * (#) added missing defer free to stubbed items
  * (#) added a errdefer deinit to oats session readLine function
  * (#) added errdefer to EVERYWHERE that returns a value owned by the caller

Recently I discovered yet another memory leak bug where, because of
how errors are handled, if a function is returning a value that is
dynamically allocated and owned by the caller, and that function
encouters an error, that function must deallocate that memory before
returning the error.

So, because of that, for the uesr interrupts and other errors that may
arise in readLine, the line memory is leaked when the error is returned
as the line.

There are two ways to fix this, either I could include a little catch
that deallocates the line upon every possible error.
ORRRRR I could let zig do that automatically for me with it's errdefer
feature.

Also, there was another memory leak error, where stubbed items didn't
actually get deallocated, so I've added that too now.

Most of this commit is just me inserting an errdefer everywhere that we
encoutner a function that returns a value that is owned by the caller.
As, previously, these resources would not be deallocated if they weren't returned (upon an err).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>